Need to figure out marking promises as complete.
When promise `i` is issued by partition `A` of acceptors, and accepted by partition `B` of acceptors; there is no way to know that the proposal has been completed on any given node in `A`.

The acceptor isn't supposed to track that information. The fact that we're satisfying a quorum says any requests to modify a record after the first modification should fail. If any acceptor has a record of writing that record then any future requests to modify it should fail. This means those requests should fail at the _proposer_ if any acceptor indicates there was an earlier request to modify the record.

They will do this by returning the promise of the earlier request. In the case the request has the same value as the current request, we can just update the N for the current request and proceed to return success to the client. If the values do not match, the request should simply be rejected at the promise phase.